+++ src/pages/profile/views/profileTabs/CreateProject.tsx
@@ -1,8 +1,8 @@
 import { useTranslation } from 'react-i18next'
 
-import { CardLayout, CardLayoutProps } from '../../../components/layouts'
-import { Body2, H2 } from '../../../components/typography'
-import { CreateAProjectButton } from '../components'
+import { CardLayout, CardLayoutProps } from '../../../../components/layouts'
+import { Body2, H2 } from '../../../../components/typography'
+import { CreateAProjectButton } from '../../components'
 
 export const CreateProject = (props: CardLayoutProps) => {
   const { t } = useTranslation()